+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.2] - 2026-05-27
 
 ### Changed
 
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Compute the correct path to the LSP server binary on Windows in the extension.
+- Compute the correct path to the LSP server binary in the Windows extension.
 
 ## [0.1.1] - 2026-05-26
 
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First pre-release of the LSP to be published to [crates.io](https://crates.io/).
 
-[unreleased]: https://github.com/0xdea/zed-highlight/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/0xdea/zed-highlight/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/0xdea/zed-highlight/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/0xdea/zed-highlight/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/0xdea/zed-highlight/releases/tag/v0.1.0